+++ static/js/forms/index.js
@@ -1,7 +1,9 @@
 import a11y from "./a11y";
 import conditional from "./conditional";
+import search from "./search";
 
 export default {
   a11y,
   conditional,
+  search,
 };
